move catalog modal to new file
